make payment and final setup

+++ src/services/BorrowerService.ts
@@ -1,10 +1,12 @@
 import { IBorrowerRepository } from '../repositories/IBorrowerRepository';
 import { IBorrower } from '../models/Borrower';
 import { CreateBorrowerDTO } from '../dto/CreateBorrowerDto';
+import { ILoanRepository } from '../repositories/ILoanRepository';
 
 export class BorrowerService {
   constructor(
-   private borrowerRepository: IBorrowerRepository
+   private borrowerRepository: IBorrowerRepository,
+   private loanRepository: ILoanRepository
   ) {}
 
   async createBorrower(data: CreateBorrowerDTO): Promise<IBorrower> {
@@ -24,6 +26,14 @@ export class BorrowerService {
   }
 
   async deleteBorrower(id: string) {
+    // Check the count of loans associated with this borrower
+    const loanCount = await this.loanRepository.countByBorrowerId(id);
+
+    if (loanCount > 0) {
+      // If there are associated loans, prevent deletion and throw an error
+      throw new Error('Cannot delete borrower. The borrower is linked with existing loans.');
+    }
+
     return await this.borrowerRepository.delete(id);
   }
 }
